RTC_ID:975329
COMMENT:[Map]同步新代码

PatchModule:Navi

+++ service/src/main/java/com/fy/navi/service/define/route/RouteL2Data.java
@@ -26,7 +26,7 @@ public class RouteL2Data {
     @SerializedName("links")
     private List<LinksDTO> mLinks;
     @SerializedName("pathID")
-    private Integer mPathID;
+    private Long mPathID;
     @SerializedName("pntCnt")
     private Integer mPntCnt;
     @SerializedName("pnts")
